refactor(executor): split local backend core
